requirements.txt
Django==5.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.1.0
python-dateutil==2.9.0.post0
whitenoise==6.8.2
gunicorn==23.0.0
coincurve==16.0.0
# Cryptography (for Stacks wallet verification)
pycryptodome==3.20.0

# Database drivers
psycopg2-binary==2.9.9  # For PostgreSQL (Railway)
dj-database-url==2.2.0  # For parsing DATABASE_URL

# Optional but recommended
python-dotenv==1.0.0  # For loading environment variables from .env file
requests==2.32.3  # For DCPE proxy HTTP calls
celery==5.4.0
redis==5.0.1
